[support] Split LogLevel into its own file
It's starting to accumulate conformances that are distracting from the
rest of the logging code.
